traders_core/requirements.core.txt
pandas>=2.1
numpy>=1.25
scikit-learn>=1.4
pydantic>=2.6
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0
MetaTrader5>=5.0.45
PyYAML>=6.0
joblib>=1.3
APScheduler>=3.10
ccxt>=4.3
prometheus-client>=0.20
scikit-learn>=1.4
